refactor treenameinput
